Fix type inconsistencies in list memory delta tracking
Change AddReportedMemorySizeDelta parameter and ReportedMemorySizeDelta
return type from size_t/int64_t to int32_t to match the underlying field
and avoid implicit narrowing at all call sites. Add explicit casts at
Upload and SetExternal. Remove dead fetches_before assignment in test.

+++ src/core/qlist.cc
@@ -430,7 +430,7 @@ void QList::Node::Upload(QList* ql, std::string_view val) {
   entry = static_cast<unsigned char*>(zmalloc(val.size()));
   memcpy(entry, val.data(), val.size());
   ql->AdjustMallocSize(val.size());
-  ql->AddReportedMemorySizeDelta(val.size());
+  ql->AddReportedMemorySizeDelta(static_cast<int32_t>(val.size()));
   ql->AdjustOffloadNodeCount(-1);
   offloaded = 0;
 }
